performed code inspection

+++ src/megan/viewer/TaxonomyData.java
@@ -205,7 +205,7 @@ public class TaxonomyData {
             {
                 Node w = v;
                 while (true) {
-                    if (!majorRanksOnly || TaxonomicLevels.isMajorRank(taxonomyClassification.getId2Rank().get((Integer) w.getInfo())))
+                    if (!majorRanksOnly || TaxonomicLevels.isMajorRank(taxonomyClassification.getId2Rank().get(w.getInfo())))
                         path.push(w);
                     if (w.getInDegree() > 0)
                         w = w.getFirstInEdge().getSource();
@@ -219,7 +219,7 @@ public class TaxonomyData {
                 Integer id = (Integer) w.getInfo();
                 if (id != null) {
                     if (majorRanksOnly) {
-                        String letters = TaxonomicLevels.getName(taxonomyClassification.getId2Rank().get((Integer) w.getInfo()));
+                        String letters = TaxonomicLevels.getName(taxonomyClassification.getId2Rank().get(w.getInfo()));
 
                         final char key = Character.toUpperCase(letters.charAt(0));
                         while (expectedIndex < expectedPath.length() && key != expectedPath.charAt(expectedIndex)) {
@@ -271,7 +271,7 @@ public class TaxonomyData {
     public static void setDisabledInternalTaxa(Set<Integer> internal) {
         final Node root = taxonomyClassification.getFullTree().getRoot();
         if (root != null)
-            setDisabledInternalTaxaRec(root, internal, internal.contains((int) root.getInfo()));
+            setDisabledInternalTaxaRec(root, internal, internal.contains(root.getInfo()));
     }
 
     /**
